Refactor requester app file names according to PEP 8
