+++ request_app/src/setup.py
@@ -1,5 +1,5 @@
-from services.RestService import RestService
-from services.GrpcService import GrpcService
+from services.rest_service import RestService
+from services.grpc_service import GrpcService
 
 restService = RestService('http://localhost:5001')
 grpcService = GrpcService('localhost:50051')
